[Refactor] - 토큰 최적화 구조 수정
