+++ build.gradle
@@ -68,10 +68,6 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 
-	// MeterRegistry
-	implementation 'org.springframework.boot:spring-boot-starter-actuator'
-	implementation 'io.micrometer:micrometer-core'
-
 	// JTokkit
 	implementation 'com.knuddels:jtokkit:1.1.0'
